clean up and remove redundant drq example
