+++ examples/async_drq_sim/async_drq_sim.py
@@ -11,6 +11,7 @@ from flax.training import checkpoints
 import cv2
 import os
 
+from typing import Any, Dict, Optional
 import pickle as pkl
 import gym
 from gym.wrappers.record_episode_statistics import RecordEpisodeStatistics
@@ -19,29 +20,25 @@ from serl_launcher.agents.continuous.drq import DrQAgent
 from serl_launcher.common.evaluation import evaluate
 from serl_launcher.utils.timer_utils import Timer
 from serl_launcher.wrappers.chunking import ChunkingWrapper
+from serl_launcher.utils.train_utils import concat_batches
 
 from agentlace.trainer import TrainerServer, TrainerClient
 from agentlace.data.data_store import QueuedDataStore
 
+from serl_launcher.data.data_store import MemoryEfficientReplayBufferDataStore
 from serl_launcher.utils.launcher import (
     make_drq_agent,
     make_trainer_config,
     make_wandb_logger,
     make_replay_buffer,
 )
-from serl_launcher.wrappers.serl_obs_wrappers import SERLObsWrapper, ManipulatorEnvObsWrapper
+from serl_launcher.wrappers.serl_obs_wrappers import SERLObsWrapper
 
 import franka_sim
 
-# new manipulator gym env # TODO remove this
-from manipulator_gym.manipulator_env import ManipulatorEnv, StateEncoding
-from manipulator_gym.interfaces.interface_service import ActionClientInterface
-from manipulator_gym.interfaces.base_interface import ManipulatorInterface
-from manipulator_gym.utils.gym_wrappers import ResizeObsImageWrapper
-
 FLAGS = flags.FLAGS
 
-flags.DEFINE_string("env", "HalfCheetah-v4", "Name of environment.")
+flags.DEFINE_string("env", "PandaPickCubeVision-v0", "Name of environment.")
 flags.DEFINE_string("agent", "drq", "Name of agent.")
 flags.DEFINE_string("exp_name", None, "Name of the experiment for wandb logging.")
 flags.DEFINE_integer("max_traj_length", 1000, "Maximum length of trajectory.")
@@ -68,9 +65,9 @@ flags.DEFINE_boolean("render", False, "Render the environment.")
 flags.DEFINE_string("ip", "localhost", "IP address of the learner.")
 # "small" is a 4 layer convnet, "resnet" and "mobilenet" are frozen with pretrained weights
 flags.DEFINE_string("encoder_type", "resnet-pretrained", "Encoder type.")
+flags.DEFINE_string("demo_path", None, "Path to the demo data.")
 flags.DEFINE_integer("checkpoint_period", 0, "Period to save checkpoints.")
 flags.DEFINE_string("checkpoint_path", None, "Path to save checkpoints.")
-flags.DEFINE_string("demo_path", None, "Path to the demo data.")
 
 flags.DEFINE_boolean(
     "debug", False, "Debug mode."
@@ -110,11 +107,11 @@ def actor(agent: DrQAgent, data_store, env, sampling_rng):
 
     client.recv_network_callback(update_params)
 
-    # eval_env = gym.make(FLAGS.env)
-    # if FLAGS.env == "PandaPickCubeVision-v0":
-    #     eval_env = SERLObsWrapper(eval_env)
-    #     eval_env = ChunkingWrapper(eval_env, obs_horizon=1, act_exec_horizon=None)
-    # eval_env = RecordEpisodeStatistics(env)
+    eval_env = gym.make(FLAGS.env)
+    if FLAGS.env == "PandaPickCubeVision-v0":
+        eval_env = SERLObsWrapper(eval_env)
+        eval_env = ChunkingWrapper(eval_env, obs_horizon=1, act_exec_horizon=None)
+    eval_env = RecordEpisodeStatistics(eval_env)
 
     obs, _ = env.reset()
     done = False
@@ -163,15 +160,15 @@ def actor(agent: DrQAgent, data_store, env, sampling_rng):
         if step % FLAGS.steps_per_update == 0:
             client.update()
 
-        # if step % FLAGS.eval_period == 0:
-        #     with timer.context("eval"):
-        #         evaluate_info = evaluate(
-        #             policy_fn=partial(agent.sample_actions, argmax=True),
-        #             env=eval_env,
-        #             num_episodes=FLAGS.eval_n_trajs,
-        #         )
-        #     stats = {"eval": evaluate_info}
-        #     client.request("send-stats", stats)
+        if step % FLAGS.eval_period == 0:
+            with timer.context("eval"):
+                evaluate_info = evaluate(
+                    policy_fn=partial(agent.sample_actions, argmax=True),
+                    env=eval_env,
+                    num_episodes=FLAGS.eval_n_trajs,
+                )
+            stats = {"eval": evaluate_info}
+            client.request("send-stats", stats)
 
         timer.tock("total")
 
@@ -183,7 +180,13 @@ def actor(agent: DrQAgent, data_store, env, sampling_rng):
 ##############################################################################
 
 
-def learner(rng, agent: DrQAgent, replay_buffer, wandb_logger=None):
+def learner(
+    rng,
+    agent: DrQAgent,
+    replay_buffer: MemoryEfficientReplayBufferDataStore,
+    demo_buffer: Optional[MemoryEfficientReplayBufferDataStore] = None,
+    wandb_logger=None,
+):
     """
     The learner loop, which runs when "--learner" is set to True.
     """
@@ -220,14 +223,32 @@ def learner(rng, agent: DrQAgent, replay_buffer, wandb_logger=None):
     server.publish_network(agent.state.params)
     print_green("sent initial network to actor")
 
+    # 50/50 sampling from RLPD, half from demo and half from online experience if
+    # demo_buffer is provided
+    demo_iterator = None
+    if demo_buffer is None:
+        single_buffer_batch_size = FLAGS.batch_size
+    else:
+        single_buffer_batch_size = FLAGS.batch_size // 2
+
     # create replay buffer iterator
     replay_iterator = replay_buffer.get_iterator(
         sample_args={
-            "batch_size": FLAGS.batch_size,
+            "batch_size": single_buffer_batch_size,
             "pack_obs_and_next_obs": True,
         },
         device=sharding.replicate(),
     )
+
+    # if demo_buffer is provided, create demo buffer iterator
+    if demo_buffer is not None:
+        demo_iterator = demo_buffer.get_iterator(
+            sample_args={
+                "batch_size": single_buffer_batch_size,
+                "pack_obs_and_next_obs": True,
+            },
+            device=sharding.replicate(),
+        )
 
     # wait till the replay buffer is filled with enough data
     timer = Timer()
@@ -238,6 +259,12 @@ def learner(rng, agent: DrQAgent, replay_buffer, wandb_logger=None):
             with timer.context("sample_replay_buffer"):
                 batch = next(replay_iterator)
 
+                # we will concatenate the demo data with the online data
+                # if demo_buffer is provided
+                if demo_iterator is not None:
+                    demo_batch = next(demo_iterator)
+                    batch = concat_batches(batch, demo_batch, axis=0)
+
             with timer.context("train_critics"):
                 agent, critics_info = agent.update_critics(
                     batch,
@@ -245,6 +272,12 @@ def learner(rng, agent: DrQAgent, replay_buffer, wandb_logger=None):
 
         with timer.context("train"):
             batch = next(replay_iterator)
+
+            # we will concatenate the demo data with the online data
+            # if demo_buffer is provided
+            if demo_iterator is not None:
+                demo_batch = next(demo_iterator)
+                batch = concat_batches(batch, demo_batch, axis=0)
             agent, update_info = agent.update_high_utd(batch, utd_ratio=1)
 
         # publish the updated network
@@ -270,26 +303,15 @@ def learner(rng, agent: DrQAgent, replay_buffer, wandb_logger=None):
 
 def main(_):
     assert FLAGS.batch_size % num_devices == 0
+
     # seed
     rng = jax.random.PRNGKey(FLAGS.seed)
 
     # create env and load dataset
-    if FLAGS.env == "PandaPickCubeVision-v0":
-        print("Using PandaPickCubeVision-v0")
-        if FLAGS.render:
-            env = gym.make(FLAGS.env, render_mode="human")
-        else:
-            env = gym.make(FLAGS.env)
+    if FLAGS.render:
+        env = gym.make(FLAGS.env, render_mode="human")
     else:
-        print("Using manipulator gym env")
-        env = ManipulatorEnv(
-            # manipulator_interface=ActionClientInterface(host=FLAGS.ip),
-            manipulator_interface=ManipulatorInterface(), # for testing
-            state_encoding=StateEncoding.POS_EULER,
-            use_wrist_cam=True,
-        )
-        env = ResizeObsImageWrapper(env, resize_size={"image_primary": (128, 128), "image_wrist": (128, 128)})
-        env = ManipulatorEnvObsWrapper(env)
+        env = gym.make(FLAGS.env)
 
     if FLAGS.env == "PandaPickCube-v0":
         env = gym.wrappers.FlattenObservation(env)
@@ -297,15 +319,7 @@ def main(_):
         env = SERLObsWrapper(env)
         env = ChunkingWrapper(env, obs_horizon=1, act_exec_horizon=None)
 
-    print(f"observation_space: {env.observation_space}")
-    print(f"action_space: {env.action_space}")
-
     image_keys = [key for key in env.observation_space.keys() if key != "state"]
-    print(f"image_keys: {image_keys}")
-    obs, _ =  env.reset()
-    for key in obs.keys():
-        print(f"key: {key}, shape: {obs[key].shape}")
-    print("\033[91m {}\033[00m".format("-"*50))
 
     rng, sampling_rng = jax.random.split(rng)
     agent: DrQAgent = make_drq_agent(
@@ -322,39 +336,10 @@ def main(_):
         jax.tree_map(jnp.array, agent), sharding.replicate()
     )
 
-
-    # NOTE: this assumes that all data are expert demonstrations
-    def preload_data_transform(data, metadata):
-        obs = data["observations"]
-        next_obs = data["next_observations"]
-        
-        # convert data img shape from (h, w, c) to (1, h, w, c)
-        # and resize to 128x128
-        for key in image_keys:
-            if len(obs[key].shape) == 3:
-                obs[key] = cv2.resize(obs[key], (128, 128))
-                obs[key] = np.expand_dims(obs[key], axis=0)
-
-            if len(next_obs[key].shape) == 3:
-                next_obs[key] = cv2.resize(next_obs[key], (128, 128))
-                next_obs[key] = np.expand_dims(next_obs[key], axis=0)
-        
-        # convert data state shape from (N,) to (1, N)
-        if len(obs["state"].shape) == 1:
-            obs["state"] = np.expand_dims(obs["state"], axis=0)
-        if len(next_obs["state"].shape) == 1:
-            next_obs["state"] = np.expand_dims(next_obs["state"], axis=0)
-
-        data["observations"] = obs
-        data["next_observations"] = next_obs
-        
-        # relabel the reward to 1 when terminated or truncated
-        # and 0 otherwise
-        if metadata["step"] == metadata["step_size"] - 1:
-            data["rewards"] = 1.0
-        else:
-            data["rewards"] = 0.0
-
+    def preload_data_transform(data, metadata) -> Optional[Dict[str, Any]]:
+        # NOTE: Create your own custom data transform function here if you
+        # are loading this via with --preload_rlds_path with tf rlds data
+        # This default does nothing
         return data
 
     def create_replay_buffer_and_wandb_logger():
@@ -383,13 +368,26 @@ def main(_):
         print_green("replay buffer created")
         print_green(f"replay_buffer size: {len(replay_buffer)}")
 
+        # if demo data is provided, load it into the demo buffer
+        # in the learner node
         if FLAGS.demo_path:
             # Check if the file exists
             if not os.path.exists(FLAGS.demo_path):
                 raise FileNotFoundError(f"File {FLAGS.demo_path} not found")
 
+            demo_buffer = MemoryEfficientReplayBufferDataStore(
+                env.observation_space,
+                env.action_space,
+                capacity=10000,
+                image_keys=image_keys,
+            )
             with open(FLAGS.demo_path, "rb") as f:
                 trajs = pkl.load(f)
+                for traj in trajs:
+                    demo_buffer.insert(traj)
+            print(f"demo buffer size: {len(demo_buffer)}")
+        else:
+            demo_buffer = None
 
         # learner loop
         print_green("starting learner loop")
@@ -397,6 +395,7 @@ def main(_):
             sampling_rng,
             agent,
             replay_buffer,
+            demo_buffer=demo_buffer,  # None if no demo data is provided
             wandb_logger=wandb_logger,
         )
