+++ serl_launcher/requirements.txt
@@ -11,7 +11,8 @@ absl-py >= 0.12.0
 scipy <= 1.12.0
 wandb >= 0.12.14
 tensorflow>=2.16.0
-tensorflow_probability>=0.23.0
+tensorflow_probability>=0.24.0
+tf_keras
 einops >= 0.6.1
 imageio >= 2.31.1
 moviepy >= 1.0.3
